requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
boto3==1.34.122
pydantic==2.7.1
python-jose[cryptography]==3.3.0
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
authlib==1.3.0